Code tidy

+++ src/test/util/bkd/test_bkd.rs
@@ -135,9 +135,9 @@ fn test_random_ints_n_dims() -> Result<()> {
     let mut random = random();
     let num_docs = at_least(&mut random, 1000);
     let dir = new_directory_shared(&mut random)?;
-    let num_dims = TestUtil::next_usize(&mut random, 1, 5) as usize;
-    let num_index_dims = TestUtil::next_usize(&mut random, 1, num_dims) as usize;
-    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 50, 100) as usize;
+    let num_dims = TestUtil::next_usize(&mut random, 1, 5);
+    let num_index_dims = TestUtil::next_usize(&mut random, 1, num_dims);
+    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 50, 100);
     let max_mb: f32 = 3.0 + (3.0 * random.random::<f32>());
     let config = BKDConfig::new(num_dims, num_index_dims, 4, max_points_in_leaf_node)?;
     let mut writer = BKDWriter::new(
@@ -270,9 +270,9 @@ fn test_big_int_n_dims() -> Result<()> {
     let num_docs = at_least(&mut random, 1000);
     let dir = new_directory_shared(&mut random)?;
 
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30) as usize;
-    let num_dims = TestUtil::next_usize(&mut random, 1, 5) as usize;
-    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 50, 100) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30);
+    let num_dims = TestUtil::next_usize(&mut random, 1, 5);
+    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 50, 100);
     let max_mb: f32 = 3.0 + (3.0 * random.random::<f32>());
     let config = BKDConfig::new(
         num_dims,
@@ -465,8 +465,8 @@ fn do_test_random_binary<R: Rng + ?Sized>(random: &mut R, count: usize) -> Resul
 fn test_all_equal() -> Result<()> {
     let mut random = random();
 
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30) as usize;
-    let num_data_dims = TestUtil::next_usize(&mut random, 1, MAX_DIMENSIONS) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30);
+    let num_data_dims = TestUtil::next_usize(&mut random, 1, MAX_DIMENSIONS);
     let num_index_dims = std::cmp::min(
         TestUtil::next_usize(&mut random, 1, num_data_dims),
         MAX_INDEX_DIMENSIONS,
@@ -503,8 +503,8 @@ fn test_all_equal() -> Result<()> {
 fn test_index_dim_equal_data_dim_different() -> Result<()> {
     let mut random = random();
 
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30) as usize;
-    let num_data_dims = TestUtil::next_usize(&mut random, 2, MAX_DIMENSIONS) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30);
+    let num_data_dims = TestUtil::next_usize(&mut random, 2, MAX_DIMENSIONS);
     let num_index_dims = std::cmp::min(
         TestUtil::next_usize(&mut random, 1, num_data_dims - 1),
         MAX_INDEX_DIMENSIONS,
@@ -552,8 +552,8 @@ fn test_index_dim_equal_data_dim_different() -> Result<()> {
 fn test_one_dim_equal() -> Result<()> {
     let mut random = random();
 
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30) as usize;
-    let num_data_dims = TestUtil::next_usize(&mut random, 1, MAX_DIMENSIONS) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30);
+    let num_data_dims = TestUtil::next_usize(&mut random, 1, MAX_DIMENSIONS);
     let num_index_dims = std::cmp::min(
         TestUtil::next_usize(&mut random, 1, num_data_dims),
         MAX_INDEX_DIMENSIONS,
@@ -577,7 +577,7 @@ fn test_one_dim_equal() -> Result<()> {
         }
     }
 
-    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 20, 50) as usize;
+    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 20, 50);
 
     verify_full(
         &mut random,
@@ -594,8 +594,8 @@ fn test_one_dim_equal() -> Result<()> {
 fn test_one_dim_low_card() -> Result<()> {
     let mut random = random();
 
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30) as usize;
-    let num_data_dims = TestUtil::next_usize(&mut random, 2, MAX_DIMENSIONS) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30);
+    let num_data_dims = TestUtil::next_usize(&mut random, 2, MAX_DIMENSIONS);
     let num_index_dims = std::cmp::min(
         TestUtil::next_usize(&mut random, 2, num_data_dims),
         MAX_INDEX_DIMENSIONS,
@@ -638,7 +638,7 @@ fn test_one_dim_low_card() -> Result<()> {
         }
     }
 
-    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 20, 50) as usize;
+    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 20, 50);
     verify_full(
         &mut random,
         &doc_values,
@@ -654,8 +654,8 @@ fn test_one_dim_low_card() -> Result<()> {
 fn test_one_dim_two_values() -> Result<()> {
     let mut random = random();
 
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30) as usize;
-    let num_data_dims = TestUtil::next_usize(&mut random, 1, MAX_DIMENSIONS) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30);
+    let num_data_dims = TestUtil::next_usize(&mut random, 1, MAX_DIMENSIONS);
     let num_index_dims = std::cmp::min(
         TestUtil::next_usize(&mut random, 1, num_data_dims),
         MAX_INDEX_DIMENSIONS,
@@ -706,8 +706,8 @@ fn test_one_dim_two_values() -> Result<()> {
 fn test_random_few_different_values() -> Result<()> {
     let mut random = random();
 
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30) as usize;
-    let num_data_dims = TestUtil::next_usize(&mut random, 1, MAX_DIMENSIONS) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30);
+    let num_data_dims = TestUtil::next_usize(&mut random, 1, MAX_DIMENSIONS);
     let num_index_dims = std::cmp::min(
         TestUtil::next_usize(&mut random, 1, num_data_dims),
         MAX_INDEX_DIMENSIONS,
@@ -758,8 +758,8 @@ impl DocMap for DocMapMock {
 fn test_multi_valued() -> Result<()> {
     let mut random = random();
 
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30) as usize;
-    let num_data_dims = TestUtil::next_usize(&mut random, 1, MAX_DIMENSIONS) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 2, 30);
+    let num_data_dims = TestUtil::next_usize(&mut random, 1, MAX_DIMENSIONS);
     let num_index_dims = std::cmp::min(
         TestUtil::next_usize(&mut random, 1, num_data_dims),
         MAX_INDEX_DIMENSIONS,
@@ -1463,11 +1463,11 @@ fn test_check_data_dim_optimal_order() -> Result<()> {
     let mut random = random();
     let dir = new_directory_shared(&mut random)?;
     let num_values = at_least(&mut random, 5000);
-    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 50, 500) as usize;
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 1, 4) as usize;
+    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 50, 500);
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 1, 4);
     let max_mb = 3.0 + 3.0 * random.random::<f64>();
-    let num_index_dims = TestUtil::next_usize(&mut random, 1, 8) as usize;
-    let num_data_dims = TestUtil::next_usize(&mut random, num_index_dims, 8) as usize;
+    let num_index_dims = TestUtil::next_usize(&mut random, 1, 8);
+    let num_data_dims = TestUtil::next_usize(&mut random, num_index_dims, 8);
 
     let mut point_value1 = vec![0u8; (num_data_dims * num_bytes_per_dim) as usize];
     let mut point_value2 = vec![0u8; (num_data_dims * num_bytes_per_dim) as usize];
@@ -1644,10 +1644,10 @@ impl IntersectVisitor for IntersectVisitorMock5<'_> {
 #[test]
 fn test_wasted_leading_bytes() -> Result<()> {
     let mut random = random();
-    let num_dims = TestUtil::next_usize(&mut random, 1, MAX_INDEX_DIMENSIONS) as usize;
-    let num_index_dims = TestUtil::next_usize(&mut random, 1, num_dims) as usize;
+    let num_dims = TestUtil::next_usize(&mut random, 1, MAX_INDEX_DIMENSIONS);
+    let num_index_dims = TestUtil::next_usize(&mut random, 1, num_dims);
     let bytes_per_dim = MAX_NUM_BYTES;
-    let bytes_used = TestUtil::next_usize(&mut random, 1, 3) as usize;
+    let bytes_used = TestUtil::next_usize(&mut random, 1, 3);
 
     let dir = new_directory_shared(&mut random)?;
     let num_docs = at_least(&mut random, 10000);
@@ -1761,8 +1761,8 @@ fn test_estimate_point_count() -> Result<()> {
     let mut random = random();
     let dir = new_directory_shared(&mut random)?;
     let num_values = at_least(&mut random, 10_000);
-    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 50, 500) as usize;
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 1, 4) as usize;
+    let max_points_in_leaf_node = TestUtil::next_usize(&mut random, 50, 500);
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 1, 4);
 
     let mut point_value = vec![0u8; num_bytes_per_dim as usize];
     let mut unique_point_value = vec![0u8; num_bytes_per_dim as usize];
@@ -1901,7 +1901,7 @@ fn test_total_point_count_validation() -> Result<()> {
     let dir = new_directory_shared(&mut random)?;
     let num_values = 10;
     let num_points_added = 50;
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 1, 4) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 1, 4);
     let mut point_value = vec![0u8; num_bytes_per_dim as usize];
     random.fill_bytes(&mut point_value);
 
@@ -2008,7 +2008,7 @@ fn test_too_many_points() -> Result<()> {
     let dir = new_directory(&mut random)?;
 
     let num_values = 10;
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 1, 4) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 1, 4);
     let mut point_value = vec![0u8; num_bytes_per_dim];
 
     let mut w = BKDWriter::new(
@@ -2046,7 +2046,7 @@ fn test_too_many_points_1d() -> Result<()> {
     let dir = new_directory_shared(&mut random)?;
 
     let num_values = 10;
-    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 1, 4) as usize;
+    let num_bytes_per_dim = TestUtil::next_usize(&mut random, 1, 4);
     let mut point_values = vec![vec![0u8; num_bytes_per_dim as usize]; 11];
     let mut doc_ids = vec![0i32; 11];
